Pull out query.rs
