+++ packages/std/src/query.rs
@@ -1,0 +1,28 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+use crate::encoding::Binary;
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum QueryResult {
+    Ok(Binary),
+    Err(String),
+}
+
+impl QueryResult {
+    // unwrap will panic on err, or give us the real data useful for tests
+    pub fn unwrap(self) -> Binary {
+        match self {
+            QueryResult::Err(msg) => panic!("Unexpected error: {}", msg),
+            QueryResult::Ok(res) => res,
+        }
+    }
+
+    pub fn is_err(&self) -> bool {
+        match self {
+            QueryResult::Err(_) => true,
+            _ => false,
+        }
+    }
+}
