+++ packages/std/src/types.rs
@@ -94,30 +94,6 @@ pub struct Coin {
     pub amount: String,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "lowercase")]
-pub enum QueryResult {
-    Ok(Binary),
-    Err(String),
-}
-
-impl QueryResult {
-    // unwrap will panic on err, or give us the real data useful for tests
-    pub fn unwrap(self) -> Binary {
-        match self {
-            QueryResult::Err(msg) => panic!("Unexpected error: {}", msg),
-            QueryResult::Ok(res) => res,
-        }
-    }
-
-    pub fn is_err(&self) -> bool {
-        match self {
-            QueryResult::Err(_) => true,
-            _ => false,
-        }
-    }
-}
-
 // coin is a shortcut constructor for a set of one denomination of coins
 pub fn coin(amount: &str, denom: &str) -> Vec<Coin> {
     vec![Coin {
